### 2023-02-05 #5130 @editorialbot
[PRE REVIEW]: geocmeans: An R package for spatial fuzzy c-means
**Submitting author:** @JeremyGelb (<a href="http://orcid.org/0000-0002-7114-2714">Jeremy Gelb</a>)
**Repository:** https://github.com/JeremyGelb/geocmeans
**Branch with paper.md** (empty if default branch):
**Version:** 0.3.2
