### 2023-02-05 #5130 @arfon
Note this is a resubmission of a previous rejected one: https://github.com/openjournals/joss-reviews/issues/3207
