### 2023-02-05 fdbbe2a
submitting 0.3.3 to CRAN and correcting references for JOSS publication
