+++ JOSS/references.bib
@@ -84,6 +84,7 @@
   volume={85},
   pages={96--103},
   year={2015},
+  doi={10.1016/j.cageo.2015.05.019},
   publisher={Elsevier}
 }
 
@@ -95,6 +96,7 @@
   number={8},
   pages={1750--1760},
   year={2011},
+  doi={10.1016/j.patcog.2011.02.009},
   publisher={Elsevier}
 }
 
@@ -106,5 +108,6 @@
   number={11},
   pages={657--664},
   year={1991},
+  doi={10.1016/0167-8655(91)90002-4},
   publisher={Elsevier}
 }
