### 2023-02-23 #5130 @Yingjie4Science
@martinfleis I might not be the right person to review this package, given the required in-depth knowledge of GIS and spatial statistics. I have sent you an email with two recommendations.
