### 2023-02-24 #5130 @nuest
This looks like an interesting submission - I'm also not an expert in the statistics, but would be willing to help out with a review.

@appelmar - have you reviewed for JOSS before? This might be right up your alley.
