### 2023-05-17 f6a4aa3
adding an error message when raster objects are given instead of SpatRast
