### 2023-06-22 #5259 @ljwolf
Hi @JeremyGelb, it looks fine to me!

Although, I think the vignette would benefit from using `sf` classes instead of sp, especially since the class of `LyonIris` will depend on the users' configuration. My default configuration read that as an `sf` rather than an `sp` object, so the vignette needed some small changes to execute. But, for users with `sp` as their default, this should run correctly.
