### 2023-06-27 #5259 @JeremyGelb
Hi @ljwolf!

Thank you for completing the review! Could you please indicated me where did you find code samples using `sp` classes instead of `sf`? I removed `sp`, `rgdal` and `maptools` from `geocmeans` dependencies when passing at version 0.2.3 but it is not impossible that I missed something in the documentation.
