### 2023-06-28 ddac175
editing the joss paper, an example with sp data classes was still here
