### 2023-06-28 #5259 @JeremyGelb
I modified the example in the paper and uploaded the new file on github. I also modified a typo in the first figure.
