### 2023-08-13 JeremyGelb/geocmeans#6 @Naeemkh
Examples Lack Seed Values, Leading to Inconsistent Results
I've noticed that the examples provided do not include seed values when utilizing random processes, such as clustering. I was wondering if you expect reproducibility or not. For example, in the following examples:

https://jeremygelb.github.io/geocmeans/articles/introduction.html

While it's understood that in clustering, some variability can be acceptable, setting a seed ensures reproducibility and consistency in the results, especially for examples and tutorials.

Thank you for your hard work on this package, and I look forward to any updates.

This issue is related to https://github.com/openjournals/joss-reviews/issues/5259 .
