### 2023-08-22 #5259 @JeremyGelb
Dear @Naeemkh ,

indeed, reproductibility is an important issue. Most of the functions in `geocmeans` have an argument `seed` that can be set by the user and that will ensure that the results are identical with the same value of seed. See for example the functions [CMeans()](https://jeremygelb.github.io/geocmeans/reference/CMeans.html), [GCMeans](https://jeremygelb.github.io/geocmeans/reference/GCMeans.html), [SFCMeans](https://jeremygelb.github.io/geocmeans/reference/SFCMeans.html), [SGFCMeans](https://jeremygelb.github.io/geocmeans/reference/SGFCMeans.html), [select_parameters](https://jeremygelb.github.io/geocmeans/reference/select_parameters.html), and [boot_group_validation](https://jeremygelb.github.io/geocmeans/reference/boot_group_validation.html). Basically, the `seed` parameter is used internally at the beginning of the execution with a call to the function `set.seed`. I will edit the documentation to make it clearer.
