+++ R/data_doc.R
@@ -24,7 +24,7 @@
 #'   \item{Y}{the Y coordinate of the center of the Iris (numeric)}
 #'   ...
 #' }
-#' @source \url{https://data.grandlyon.com/}
+#' @source \url{https://data.grandlyon.com/portail/fr/accueil}
 "LyonIris"
 
 
