+++ man/adj_spconsist_arr_window_globstd.Rd
@@ -11,12 +11,12 @@ adj_spconsist_arr_window_globstd(data, memberships, window, mindist = 1e-11)
 
 \item{memberships}{an arma cube of dimension nr, nc, ks}
 
+\item{window}{a matrix representing the neighbouring of each pixel}
+
 \item{mindist}{A minimum value for distance between two observations. If two
 neighbours have exactly the same values, then the euclidean distance
 between them is 0, leading to an infinite spatial weight. In that case,
 the minimum distance is used instead of 0.}
-
-\item{a}{matrix representing the neighbouring of each pixel}
 }
 \value{
 a double, the adjusted spatial inconsitency index
