### 2023-08-26 dcdf690
little changes in the doc
