### 2023-08-26 654d9fb
small edits in the documentation
