+++ vignettes/introduction.Rmd
@@ -575,6 +575,11 @@ nl_SFCM <- SFCMeans(Data, WMat_adj, k = 4, m = 1.5, alpha = 0.7,
 ```
 When working with rasters, the same method can be apllied by giving `"nl"` as value for the parameter `lag_method`.
 
+## Reproductibility
+
+Clustering methods like CMeans depend on the initial centers selected. In `geocmeans`, they are selected randomly, and two runs of the functions can yield different results. To facilitate the reproductibility of the results, the main functions of the package (`CMeans`, `GFCMeans`, `SFCMeans`, `SGFCMeans`, `selectParameters`, `selectParameters.mc`) have a `seed` parameter. It can be set by the user to ensure that the results of the functions are exactly the same.
+
+
 ## Final words
 
 That's all, folks ! Following are the enhancements for the next version
