### 2023-08-26 JeremyGelb/geocmeans#6 @JeremyGelb
A paragraph has been added here :
https://jeremygelb.github.io/geocmeans/ and here https://jeremygelb.github.io/geocmeans/articles/introduction.html
![image](https://github.com/JeremyGelb/geocmeans/assets/46856737/a4067250-de48-43e7-a4a6-01e7b8ef0278)

Feel free to reopen the issue if you think that it should be addressed in another way.
