### 2023-09-04 c588a28
minor changes in doc for publication in JOSS
