### 2023-09-05 8f2daac
minor correction in documentation and of the function is.FCMres after submission to CRAN
