### 2023-09-06 #5259 @JeremyGelb
Thank you all. I am preparing the new submission to CRAN, but as always I got weird NOTES. I will proceed with the checklist when everything is ready.

This time, I got on debian :
```
* checking re-building of vignette outputs ... [279s/99s] NOTE
Re-building vignettes had CPU time 2.8 times elapsed time
```

I have contacted the CRAN team because I could not find online what it means.
