### 2023-09-12 e36994d
removing vignettes from CRAN release. They will be only available on the website now
