+++ vignettes/introduction.R
@@ -1,343 +1,0 @@
-## ----setup, include=FALSE-----------------------------------------------------
-knitr::opts_chunk$set(echo = TRUE)
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-#charging packages and data
-library(geocmeans)
-library(ggplot2)
-library(ggpubr)
-library(dplyr)
-library(viridis)
-library(spdep)
-
-data(LyonIris)
-
-# selecting the columns for the analysis
-AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14",
-                   "Pct_65","Pct_Img","TxChom1564","Pct_brevet","NivVieMed")
-
-# rescaling the columns
-Data <- LyonIris@data[AnalysisFields]
-for (Col in names(Data)){
-  Data[[Col]] <- scale(Data[[Col]])
-}
-
-# preparing some elements for further mapping
-LyonIris$OID <- as.character(1:nrow(LyonIris))
-FortiData <- ggplot2::fortify(LyonIris,region="OID")
-
-## ----include=FALSE------------------------------------------------------------
-# loading the pre-calculated results
-load(system.file("extdata", "results_vignette_intro.rda",
-                           package = "geocmeans", mustWork = TRUE))
-
-## ----warning=FALSE, fig.cap = "Impact of the number of groups on the explained variance", out.width = "50%", fig.pos="H", fig.align="center"----
-# finding the best k by using the r2 of the classification
-# trying for k from 2 to 10
-R2s <- sapply(2:10,function(k){
-  Clust <- kmeans(Data,centers=k,iter.max = 150)
-  R2 <- Clust$betweenss / Clust$totss
-  return(R2)
-})
-
-Df <- data.frame(K=2:10,
-                 R2 = R2s)
-
-ggplot(Df)+
-  geom_line(aes(x=K,y=R2s))+
-  geom_point(aes(x=K,y=R2s),color="red")+
-  xlab("Number of groups")+
-  ylab("R2 of classification")
-
-## ----warning=FALSE, fig.cap="Clusters identified by classical k-means", fig.pos="H", fig.align="center"----
-KMeanClust <-  kmeans(Data,centers=4,iter.max = 150)
-LyonIris$Cluster <-paste("cluster",KMeanClust$cluster,sep="_")
-
-# mapping the groups
-DFmapping <- merge(FortiData,LyonIris,by.x="id",by.y="OID")
-
-ggplot(data=DFmapping)+
-  geom_polygon(aes(x=long,y=lat,group=group,fill=Cluster),color=rgb(0,0,0,0))+
-  coord_fixed(ratio = 1)+
-  scale_fill_manual(name="Cluster",values = c("cluster_1"="palegreen3",
-                                 "cluster_2"="firebrick",
-                                 "cluster_3"="lightyellow2",
-                                 "cluster_4"="steelblue"))+
-  theme( axis.title = ggplot2::element_blank(),
-    axis.text = ggplot2::element_blank(),
-    axis.ticks = ggplot2::element_blank()
-    )
-
-
-## ----warning=FALSE------------------------------------------------------------
-Cmean <- CMeans(Data,4,1.5,500,standardize = FALSE, seed = 456, tol = 0.00001, verbose = FALSE)
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-calcqualityIndexes(Data, Cmean$Belongings, m = 1.5)
-
-## ----warning=FALSE, eval = FALSE----------------------------------------------
-#  beta_values <- selectParameters("GFCM",data = Data, k = 4, m = 1.5,
[…]
