+++ vignettes/FCMres.R
@@ -1,156 +1,0 @@
-## ----setup, include=FALSE-----------------------------------------------------
-knitr::opts_chunk$set(echo = TRUE)
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-library(geocmeans)
-library(tmap)
-library(dplyr)
-library(ggplot2)
-library(spdep)
-library(raster)
-
-data("LyonIris")
-
-# selecting the columns for the analysis
-AnalysisFields <-c("Lden","NO2","PM25","VegHautPrt","Pct0_14",
-                   "Pct_65","Pct_Img","TxChom1564","Pct_brevet","NivVieMed")
-
-# rescaling the columns
-Data <- LyonIris@data[AnalysisFields]
-for (Col in names(Data)){
-  Data[[Col]] <- scale(Data[[Col]])
-}
-
-# applying the hclust function
-clust <- hclust(dist(Data), method = "ward")
-
-# getting the groups
-LyonIris$Hclust_groups <- as.character(cutree(clust, k = 4))
-Data$Hclust_groups <- as.character(cutree(clust, k = 4))
-
-# mapping the groups
-tm_shape(LyonIris) + 
-  tm_polygons(col = "Hclust_groups", title = "groups")
-
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-centers <- Data %>% 
-  group_by(Hclust_groups) %>%
-  summarise_all(mean)
-
-centers <- as.matrix(centers[2:ncol(centers)])
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-member_mat <- cat_to_belongings(Data$Hclust_groups)
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-Data$Hclust_groups <- NULL
-
-hclustres <- FCMres(list(
-  "Centers" = centers,
-  "Belongings" = member_mat,
-  "Data" = Data,
-  "m" = 1,
-  "algo" = "hclust"
-))
-
-## ----message=FALSE, warning=FALSE, eval = FALSE-------------------------------
-#  # quick summaries about the groups
-#  summary(hclustres)
-#  violinPlots(hclustres$Data, hclustres$Groups)
-#  spiderPlots(hclustres$Data, hclustres$Belongings)
-#  mapClusters(LyonIris, hclustres)
-#  
-#  # some indices about classification quality
-#  calcqualityIndexes(hclustres$Data,
-#                     hclustres$Belongings,
-#                     hclustres$m)
-#  
-#  # spatial diagnostic
-#  Neighbours <- poly2nb(LyonIris,queen = TRUE)
-#  WMat <- nb2listw(Neighbours,style="W",zero.policy = TRUE)
-#  spatialDiag(hclustres, nblistw = WMat)
-#  
-#  # investigation with the shiny app
-#  sp_clust_explorer(hclustres, spatial = LyonIris)
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-data("Arcachon")
-
-# loading each raster as a column in a matrix
-# and scale each column
-all_data <- do.call(cbind, lapply(names(Arcachon), function(n){
-  rast <- Arcachon[[n]]
-  return(raster::values(raster::scale(rast)))
-}))
-
-# removing the rows with missing values
-missing <- complete.cases(all_data)
-all_data <- all_data[missing,]
-
-# applying the kmeans algorithm with 7 groups
-kmean7 <- kmeans(all_data, 7)
-
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-
-# creating Data (do not forget the standardization)
[…]
