### 2023-09-12 #5259 @JeremyGelb
So, I managed to resubmit to CRAN,

Here is the Zenodo DOI : https://doi.org/10.5281/zenodo.8316593

The title, the authors and the ORCIDs are matching.

A release was created on github : https://github.com/JeremyGelb/geocmeans/releases/tag/documentation
