### 2023-09-14 #5259 @martinfleis
@JeremyGelb thanks. Tags usually represent a version of each release, e.g. `v0.3.4`. In you case, the tag is `documentation`. It is unclear how that is planned to evolve in future. See for example `{sf}`'s approach to [tags](https://github.com/r-spatial/sf/tags). JOSS expectations are not explicit here but the rules implicitly assume that the tag follows the pattern tag==version. Would you mind changing that? I suppose the system would be fine with the `documentation` tag as well, but it is a bit strange, so if that is not an issue, I'd prefer a standard tag.
