### 2023-09-14 #5259 @JeremyGelb
@martinfleis, I just changed the tag of the release and it is now v0.3.4 as requested.
