### 2023-09-21 #5259 @martinfleis
@JeremyGelb the paper contains a duplicated figure label `\label{fig:example}` causing the acceptance workflow to fail. Could you fix the labels please?
