### 2023-09-22 4dd7cd1
edit for publication in JOSS
