### 2023-09-22 #5259 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
